Refactor utility functions to separate client and server - Resolves #14
