+++ utils/client/index.ts
@@ -1,0 +1,5 @@
+// Export all client-side utilities for easier imports
+export * from './annotation';
+export * from './api';
+export * from './image';
+export * from './label';
